fix: add TS project
