--- packages/examples/tokens-usage/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  rules: {
-    'react/jsx-uses-react': 'off',
-    'react/react-in-jsx-scope': 'off',
-    'react/no-unstable-nested-components': 'off',
-    'no-unsafe-optional-chaining': 'off',
-    'no-nonoctal-decimal-escape': 'off',
-  },
-};
