--- packages/examples/tokens-usage/src/App.js
@@ -1,29 +1,0 @@
-import './App.css';
-import { ThemeProvider } from '@razorpay/blade/components';
-import { paymentTheme } from '@razorpay/blade/tokens';
-import logo from './logo.svg';
-
-function App() {
-  return (
-    <ThemeProvider theme={paymentTheme}>
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <p>
-            Edit <code>src/App.js</code> and save to reload.
-          </p>
-          <a
-            className="App-link"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-          </a>
-        </header>
-      </div>
-    </ThemeProvider>
-  );
-}
-
-export default App;
